Improve QoL for OSS users of FBShipIt
Summary:
This is a minor refactor that makes it a little easier to work with the public version of shipit.

I found when writing the demo that it was odd that there was no built-in filter to do this, but I realize now that's because in the OSS version it's tucked away in fb-examples and I missed it.

Differential Revision: D21642531

fbshipit-source-id: 9bf4349852f4cca0f6989c7c303735a6a4ebda34