Make entrypoints async
Summary:
Long term, ShipIt will need to be async to support HSL IO and to stop using `awaitSynchrounously` in a few places. This is the first of several diffs to async-ify ShipIt.

I converted everything in one giant, painful commit over the course of several days and then I've been working on splitting it into review-sized diffs. Knowing the end-game has some advantages for me as the author but a few changes might look odd or unnecessary in each individual diff for the reviewer.

Differential Revision: D22168831

fbshipit-source-id: 312d9cf37294ddbe979a8e0fb015e0d5b39e41b7