Support rewriting C++ include directives (#111)
Summary:
Pull Request resolved: https://github.com/facebook/fbshipit/pull/111

After some discussions with some C++ folks, I think we can improve how we handle projects that aren't hosted at the root level of the source repo.

C++ conventions at FB prefer absolute paths for `#include "x"` directives, for example in a fictional `fbsource/fbcode/some/deep/project/file.cpp`:

```
#include <functional> // a stdlib header
#include "some/deep/project/header.h"
```

The problem is that when this file is exported to GitHub we would map "some/deep/project/" to "project/" with ShipIt so that the includes in that file no longer make sense. This is what we should export:

```
#include <functional> // a stdlib header
#include "project/header.h"
```

Historically this hasn't been a problem because our C++ open sourced projects have been hosted at the root level (e.g. `fbsource/fbcode/folly/*`). However, we have a few projects that are running up against this problem and we should try and make it as easy as possible to make a project open source **and** comply with internal conventions

Reviewed By: yfeldblum, yns88

Differential Revision: D22559398

fbshipit-source-id: e00a833b77e482f45101f43576e2d3f4cadbefa6