Rename ShipItBaseConfig to ShipItManifest
Summary:
And all the variables names too.

We have too many configs in ShipIt:
1. FBShipItConfig
2. FBSourceBranchConfig
3. FBShipItCLIStaticConfig
4. ShipItBaseConfig

This is endlessly confusing for newcomers. "Manifest" means the same thing but it is visually and semantically distinct. I plan on renaming `FBSourceBranchConfig` as soon as I can think of a good name. And `FBShipItCLIStaticConfig` should just be normal fields on the config.

Reviewed By: jurajh-fb

Differential Revision: D23588313

fbshipit-source-id: ccc572e3b919c26b007cc3f6f5b5f92f86a1c435