Use full git SHAs
Summary: Right now mercurial SHAs are always full-length, but git shas are repo-dependent when used as a ShipIt source. `git log --oneline` picks the length of the SHA it renders by the number of commits. This is nice for readability/users but a headache for automation that attempts to look for the tracking SHAs that ShipIt produces

Differential Revision: D34693603

fbshipit-source-id: 0fa328a4ec4688390d783c341872e66faa20122c