Add new option to sync from a specific rev
Summary:
Obviously this is a bad idea for committing then pushing, but this is useful when "syncing" a Phabricator Diff with ShipIt in order to create a Pull Request.

A later step in the export process will create a new branch and push that to a fork, we just need to unwind the base branch back to the commit we want to export.

A later diff will make the necessary changes to the export job to translate revisions from Meta internal to GitHub, then set this option appropriately.

Differential Revision: D35943443

fbshipit-source-id: c59548245fc43206f0f487a80fab6ebf20d9b5b1