Fix verification for .gitignore and .gitattributes at the root of the source
Summary:
I can't believe we've never hit this bug before but it turns out that `git archive` has a default list of excluded files (the 2 files in the title). We can write an override file to a secret spot inside `.git` (I feel a bit quesy about this, but this is [a documented solution](https://git-scm.com/docs/git-archive#ATTRIBUTES).

This is being hit by `mysql-5.6` right now and they pinged me and asked about this. This should fix it for them (see test plan)

Reviewed By: jurajh-fb

Differential Revision: D42463780

fbshipit-source-id: b0de6a3441eb8118e02ea90cc1d603aa42132a25